requirements.txt
fastapi
uvicorn
jinja2
sqlite3
llama-cpp-python